requirements.txt
mcp[cli]
anthropic
gspread
python-dotenv
